Extract common config. to own file and always restart affected services
Configuration related to data sources (RabbitMQ and Redis) is not
actually specific to client or server, but common to both, and hence,
needs to triggers restarts for all services.

Extract such tasks to a separate file, and make sure to restart all the
services whenever they change.

This also removes the duplicate Redis config. write that previously
existed.